+++ wxc_sdk/as_api.py
@@ -60,22 +60,24 @@ __all__ = ['AsAccessCodesApi', 'AsAdminAuditEventsApi', 'AsAgentCallerIdApi', 'A
            'AsDeviceConfigurationsApi', 'AsDeviceSettingsJobsApi', 'AsDevicesApi', 'AsDialPlanApi',
            'AsDigitPatternsApi', 'AsDndApi', 'AsEventsApi', 'AsExecAssistantApi', 'AsForwardingApi', 'AsGroupsApi',
            'AsGuestManagementApi', 'AsHotelingApi', 'AsHuntGroupApi', 'AsIncomingPermissionsApi',
-           'AsInternalDialingApi', 'AsJobsApi', 'AsLicensesApi', 'AsLocationAccessCodesApi', 'AsLocationInterceptApi',
-           'AsLocationMoHApi', 'AsLocationNumbersApi', 'AsLocationVoicemailSettingsApi', 'AsLocationsApi',
-           'AsManageNumbersJobsApi', 'AsMeetingChatsApi', 'AsMeetingClosedCaptionsApi', 'AsMeetingInviteesApi',
-           'AsMeetingParticipantsApi', 'AsMeetingPreferencesApi', 'AsMeetingQandAApi', 'AsMeetingQualitiesApi',
-           'AsMeetingTranscriptsApi', 'AsMeetingsApi', 'AsMembershipApi', 'AsMessagesApi', 'AsMonitoringApi',
-           'AsNumbersApi', 'AsOrganisationVoicemailSettingsAPI', 'AsOrganizationApi', 'AsOutgoingPermissionsApi',
-           'AsPagingApi', 'AsPeopleApi', 'AsPersonForwardingApi', 'AsPersonSettingsApi', 'AsPersonSettingsApiChild',
-           'AsPreferredAnswerApi', 'AsPremisePstnApi', 'AsPrivacyApi', 'AsPrivateNetworkConnectApi',
-           'AsPushToTalkApi', 'AsRebuildPhonesJobsApi', 'AsReceptionistApi', 'AsReceptionistContactsDirectoryApi',
-           'AsRecordingsApi', 'AsReportsApi', 'AsRestSession', 'AsRoomTabsApi', 'AsRoomsApi', 'AsRouteGroupApi',
-           'AsRouteListApi', 'AsSCIM2BulkApi', 'AsSCIM2UsersApi', 'AsScheduleApi', 'AsScimApiChild', 'AsScimV2Api',
-           'AsStatusAPI', 'AsTeamMembershipsApi', 'AsTeamsApi', 'AsTelephonyApi', 'AsTelephonyDevicesApi',
-           'AsTelephonyLocationApi', 'AsTransferNumbersApi', 'AsTrunkApi', 'AsVirtualLinesApi', 'AsVoiceMessagingApi',
-           'AsVoicePortalApi', 'AsVoicemailApi', 'AsVoicemailGroupsApi', 'AsVoicemailRulesApi', 'AsWebexSimpleApi',
-           'AsWebhookApi', 'AsWorkspaceDevicesApi', 'AsWorkspaceLocationApi', 'AsWorkspaceLocationFloorApi',
-           'AsWorkspaceNumbersApi', 'AsWorkspacePersonalizationApi', 'AsWorkspaceSettingsApi', 'AsWorkspacesApi']
+           'AsInternalDialingApi', 'AsJobsApi', 'AsLicensesApi', 'AsLocationAccessCodesApi',
+           'AsLocationEmergencyServicesApi', 'AsLocationInterceptApi', 'AsLocationMoHApi', 'AsLocationNumbersApi',
+           'AsLocationVoicemailSettingsApi', 'AsLocationsApi', 'AsManageNumbersJobsApi', 'AsMeetingChatsApi',
+           'AsMeetingClosedCaptionsApi', 'AsMeetingInviteesApi', 'AsMeetingParticipantsApi',
+           'AsMeetingPreferencesApi', 'AsMeetingQandAApi', 'AsMeetingQualitiesApi', 'AsMeetingTranscriptsApi',
+           'AsMeetingsApi', 'AsMembershipApi', 'AsMessagesApi', 'AsMonitoringApi', 'AsNumbersApi',
+           'AsOrgEmergencyServicesApi', 'AsOrganisationVoicemailSettingsAPI', 'AsOrganizationApi',
+           'AsOutgoingPermissionsApi', 'AsPagingApi', 'AsPeopleApi', 'AsPersonForwardingApi', 'AsPersonSettingsApi',
+           'AsPersonSettingsApiChild', 'AsPreferredAnswerApi', 'AsPremisePstnApi', 'AsPrivacyApi',
+           'AsPrivateNetworkConnectApi', 'AsPushToTalkApi', 'AsRebuildPhonesJobsApi', 'AsReceptionistApi',
+           'AsReceptionistContactsDirectoryApi', 'AsRecordingsApi', 'AsReportsApi', 'AsRestSession', 'AsRoomTabsApi',
+           'AsRoomsApi', 'AsRouteGroupApi', 'AsRouteListApi', 'AsSCIM2BulkApi', 'AsSCIM2UsersApi', 'AsScheduleApi',
+           'AsScimApiChild', 'AsScimV2Api', 'AsStatusAPI', 'AsTeamMembershipsApi', 'AsTeamsApi', 'AsTelephonyApi',
+           'AsTelephonyDevicesApi', 'AsTelephonyLocationApi', 'AsTransferNumbersApi', 'AsTrunkApi',
+           'AsUserEmergencyServicesApi', 'AsVirtualLinesApi', 'AsVoiceMessagingApi', 'AsVoicePortalApi',
+           'AsVoicemailApi', 'AsVoicemailGroupsApi', 'AsVoicemailRulesApi', 'AsWebexSimpleApi', 'AsWebhookApi',
+           'AsWorkspaceDevicesApi', 'AsWorkspaceLocationApi', 'AsWorkspaceLocationFloorApi', 'AsWorkspaceNumbersApi',
+           'AsWorkspacePersonalizationApi', 'AsWorkspaceSettingsApi', 'AsWorkspacesApi']
 
 
 @dataclass(init=False)
@@ -5990,10 +5992,12 @@ class AsPersonSettingsApiChild(AsApiChild, base=''):
         alternates = {('workspaces', 'musicOnHold'): ('telephony/config/workspaces', '/'),
                       ('workspaces', 'outgoingPermission/digitPatterns'): ('telephony/config/workspaces', '/'),
                       ('workspaces', 'callBridge'): ('telephony/config/workspaces', '/'),
+                      ('workspaces', 'emergencyCallbackNumber'): ('telephony/config/workspaces', '/'),
                       ('people', 'outgoingPermission/'): ('telephony/config/people', '/'),
                       ('people', 'outgoingPermission/accessCodes'): ('telephony/config/people', '/'),
                       ('people', 'outgoingPermission/digitPatterns'): ('telephony/config/people', '/'),
                       ('people', 'callBridge'): ('telephony/config/people', '/features/'),
+                      ('people', 'emergencyCallbackNumber'): ('telephony/config/people', '/'),
                       }
         selector, feature_prefix = alternates.get((selector, self.feature), (selector, feature_prefix))
         return self.session.ep(f'{selector}/{person_id}{feature_prefix}{self.feature}{path}')
@@ -8071,6 +8075,121 @@ class AsScheduleApi(AsApiChild, base='telephony/config/locations'):
         await self.delete(url, params=params)
 
 
+class AsUserEmergencyServicesApi(AsPersonSettingsApiChild):
+    """
+    Emergency Callback Configurations can be enabled at the organization level, Users without individual telephone
+    numbers, such as extension-only users, must be set up with accurate Emergency Callback Numbers (ECBN) to enable
+    them to make emergency calls. These users can either utilize the default ECBN for their location or be assigned
+    another specific telephone number from that location for emergency purposes.
+
+    Viewing these settings requires a full or read-only administrator auth token with a scope of
+    `spark-admin:telephony_config_read`. Modifying these settings requires a full administrator auth token with a
+    scope of `spark-admin:telephony_config_write`.
+    """
+
+    feature = 'emergencyCallbackNumber'
+
+    async def read(self, entity_id: str,
+             org_id: str = None) -> PersonECBN:
+        """
+        Get a Person's Emergency Callback Number
+
+        Retrieve a person's emergency callback number settings. Also applies to workspaces and virtual lines.
+
+        Emergency Callback Configurations can be enabled at the organization level, Users without individual telephone
+        numbers, such as extension-only users, must be set up with accurate Emergency Callback Numbers (ECBN) and
+        Emergency Service Addresses to enable them to make emergency calls. These users can either utilize the default
+        ECBN for their location or be assigned another specific telephone number from that location for emergency
+        purposes.
+
+        To retrieve a person's callback number requires a full, user or read-only administrator or location
+        administrator auth token with a scope of `spark-admin:telephony_config_read`.
+
+        :param entity_id: Unique identifier for the person, virtual line, or workspace
+        :type entity_id: str
+        :param org_id: ID of the organization within which the person resides. Only admin users of another organization
+            (such as partners) may use this parameter as the default is the same organization as the token used to
+            access API.
+        :type org_id: str
+        :rtype: :class:`PersonECBN`
+        """
+        params = {}
+        if org_id is not None:
+            params['orgId'] = org_id
+        url = self.f_ep(entity_id)
+        data = await super().get(url, params=params)
+        r = PersonECBN.model_validate(data)
+        return r
+
+    async def configure(self, entity_id: str,
+                  selected: SelectedECBN,
+                  location_member_id: str = None, org_id: str = None):
+        """
+        Update a Person's Emergency Callback Number.
+
+        Update a person's emergency callback number settings. Also applies to workspaces and virtual lines.
+
+        Emergency Callback Configurations can be enabled at the organization level, Users without individual telephone
+        numbers, such as extension-only users, must be set up with accurate Emergency Callback Numbers (ECBN) to
+        enable them to make emergency calls. These users can either utilize the default ECBN for their location or be
+        assigned another specific telephone number from that location for emergency purposes.
+
+        To update an emergency callback number requires a full, location, user, or read-only administrator auth token
+        with a scope of `spark-admin:telephony_config_write`.
+
+        :param entity_id: Unique identifier for the person, virtual line, or workspace.
+        :type entity_id: str
+        :param selected: The source from which the emergency calling line ID (CLID) is selected for an actual emergency
+            call.
+        :type selected: SelectedECBN
+        :param location_member_id: Member ID of person/workspace/virtual line within the location.
+        :type location_member_id: str
+        :param org_id: ID of the organization within which the person resides. Only admin users of another organization
+            (such as partners) may use this parameter as the default is the same organization as the token used to
+            access API.
+        :type org_id: str
+        :rtype: None
+        """
+        params = {}
+        if org_id is not None:
+            params['orgId'] = org_id
+        body = dict()
+        body['selected'] = enum_str(selected)
+        if location_member_id is not None:
+            body['locationMemberId'] = location_member_id
+        url = self.f_ep(entity_id)
+        await super().put(url, params=params, json=body)
+
+    async def dependencies(self, entity_id: str,
+                     org_id: str = None) -> ECBNDependencies:
+        """
+        Retrieve A Person's Emergency Callback Number Dependencies
+
+        Retrieve Emergency Callback Number dependencies for a person. Also applies to workspaces and virtual lines.
+
+        Emergency Callback Configurations can be enabled at the organization level, Users without individual telephone
+        numbers, such as extension-only users, must be set up with accurate Emergency Call Back Numbers (ECBN) to
+        enable them to make emergency calls. These users can either utilize the default ECBN for their location or be
+        assigned another specific telephone number from that location for emergency purposes.
+
+        Retrieving the dependencies requires a full, user or read-only administrator or location administrator auth
+        token with a scope of `spark-admin:telephony_config_read`.
+
+        :param entity_id: Unique identifier for the person, virtual line, or workspace
+        :type entity_id: str
+        :param org_id: Retrieve Emergency Callback Number attributes for this organization.
+        :type org_id: str
+        :rtype: :class:`ECBNDependencies`
+        """
+        params = {}
+        if org_id is not None:
+            params['orgId'] = org_id
+        url = self.f_ep(entity_id, 'dependencies')
+        data = await super().get(url, params=params)
+        r = ECBNDependencies.model_validate(data)
+        return r
+
+
 class AsVoicemailApi(AsPersonSettingsApiChild):
     """
     API for person's call voicemail settings
@@ -8271,6 +8390,8 @@ class AsPersonSettingsApi(AsApiChild, base='people'):
     caller_id: AsCallerIdApi
     #: Person's Calling Behavior
     calling_behavior: AsCallingBehaviorApi
+    #: ECBN settings
+    ecbn: AsUserEmergencyServicesApi
     #: Executive Assistant Settings for a Person
     exec_assistant: AsExecAssistantApi
     #: Forwarding Settings for a Person
@@ -8303,13 +8424,14 @@ class AsPersonSettingsApi(AsApiChild, base='people'):
         self.agent_caller_id = AsAgentCallerIdApi(session=session)
         self.appservices = AsAppServicesApi(session=session)
         self.barge = AsBargeApi(session=session)
-        self.dnd = AsDndApi(session=session)
         self.call_bridge = AsCallBridgeApi(session=session)
         self.call_intercept = AsCallInterceptApi(session=session)
         self.call_recording = AsCallRecordingApi(session=session)
         self.call_waiting = AsCallWaitingApi(session=session)
         self.calling_behavior = AsCallingBehaviorApi(session=session)
         self.caller_id = AsCallerIdApi(session=session)
+        self.dnd = AsDndApi(session=session)
+        self.ecbn = AsUserEmergencyServicesApi(session=session)
         self.exec_assistant = AsExecAssistantApi(session=session)
         self.forwarding = AsPersonForwardingApi(session=session)
         self.hoteling = AsHotelingApi(session=session)
@@ -13973,6 +14095,73 @@ class AsLocationInterceptApi(AsApiChild, base='telephony/config/locations'):
         await self.put(ep, params=params, data=data)
 
 
+class AsOrgEmergencyServicesApi(AsApiChild, base='telephony/config/emergencyCallNotification'):
+    """
+    Organization Call Settings with Emergency Services
+
+    Emergency Call Notifications can be enabled at the organization level, allowing specified email addresses to
+    receive email notifications when an emergency call is made. To comply with U.S. Public Law 115-127, also known as
+    Kari’s Law, any call that's made from within your organization to emergency services must generate an email
+    notification.
+
+    Viewing these organization settings requires a full or read-only administrator auth token with a scope of
+    `spark-admin:telephony_config_read`. Modifying these organization settings requires a full administrator auth
+    token with a scope of `spark-admin:telephony_config_write`.
+    """
+
+    async def read_emergency_call_notification(self, org_id: str = None) -> OrgEmergencyCallNotification:
+        """
+        Get an Organization Emergency Call Notification
+
+        Get organization emergency call notification.
+
+        Emergency Call Notifications can be enabled at the organization level, allowing specified email addresses to
+        receive email notifications when an emergency call is made. To comply with U.S. Public Law 115-127, also known
+        as Kari’s Law, any call that's made from within your organization to emergency services must generate an email
+        notification.
+
+        To retrieve organization call notifications requires a full, user or read-only administrator auth token with a
+        scope of `spark-admin:telephony_config_read`.
+
+        :param org_id: Retrieve Emergency Call Notification attributes for the organization.
+        :type org_id: str
+        :rtype: :class:`OrgEmergencyCallNotification`
+        """
+        params = {}
+        if org_id is not None:
+            params['orgId'] = org_id
+        url = self.ep()
+        data = await super().get(url, params=params)
+        r = OrgEmergencyCallNotification.model_validate(data)
+        return r
+
+    async def update_emergency_call_notification(self, setting: OrgEmergencyCallNotification, org_id: str = None):
+        """
+        Update an organization emergency call notification.
+
+        Once settings are enabled at the organization level, the configured email address will receive emergency call
+        notifications for all locations.
+
+        Emergency Call Notifications can be enabled at the organization level, allowing specified email addresses to
+        receive email notifications when an emergency call is made. To comply with U.S. Public Law 115-127, also known
+        as Kari’s Law, any call that's made from within your organization to emergency services must generate an email
+        notification.
+
+        To update organization call notification requires a full or user administrator auth token with a scope of
+        `spark-admin:telephony_config_write`.
+
+        :param setting: updated settings
+        :type setting: OrgEmergencyCallNotification
+        :param org_id: Update Emergency Call Notification attributes for the organization.
+        :type org_id: str
+        :rtype: None
+        """
+        params = org_id and {'orgId': org_id} or None
+        body = setting.update()
+        url = self.ep()
+        await super().put(url, params=params, json=body)
+
+
 class AsOrganisationVoicemailSettingsAPI(AsApiChild, base='telephony/config/voicemail/settings'):
     """
     API for Organisation voicemail settings
@@ -16341,6 +16530,81 @@ class AsInternalDialingApi(AsApiChild, base='telephony/config/locations'):
         await self.put(url=url, params=params, data=data)
 
 
+class AsLocationEmergencyServicesApi(AsApiChild, base='telephony/config/locations'):
+    """
+    Emergency Call Notifications can be enabled at the organization level, allowing specified email addresses to
+    receive email notifications when an emergency call is made. Once activated at the organization level, individual
+    locations can configure this setting to direct notifications to specific email addresses. To comply with U.S.
+    Public Law 115-127, also known as Kari’s Law, any call that's made from within your organization to emergency
+    services must generate an email notification.
+
+    Viewing these organization settings requires a full or read-only administrator auth token with a scope
+    of `spark-admin:telephony_config_read`. Modifying these organization settings requires a full administrator auth
+    token with a scope of `spark-admin:telephony_config_write`.
+    """
+
+    async def read_emergency_call_notification(self, location_id: str,
+                                                   org_id: str = None) -> LocationEmergencyCallNotification:
+        """
+        Get a Location Emergency Call Notification
+
+        Get location emergency call notification.
+
+        Emergency Call Notifications can be enabled at the organization level, allowing specified email addresses to
+        receive email notifications when an emergency call is made. Once activated at the organization level,
+        individual locations can configure this setting to direct notifications to specific email addresses. To comply
+        with U.S. Public Law 115-127, also known as Kari’s Law, any call that's made from within your organization to
+        emergency services must generate an email notification.
+
+        To retrieve location call notifications requires a full, user or read-only administrator or location
+        administrator auth token with a scope of `spark-admin:telephony_config_read`.
+
+        :param location_id: Retrieve Emergency Call Notification attributes for this location.
+        :type location_id: str
+        :param org_id: Retrieve Emergency Call Notification attributes for the location in this organization.
+        :type org_id: str
+        :rtype: :class:`LocationEmergencyCallNotification`
+        """
+        params = {}
+        if org_id is not None:
+            params['orgId'] = org_id
+        url = self.ep(f'{location_id}/emergencyCallNotification')
+        data = await super().get(url, params=params)
+        r = LocationEmergencyCallNotification.model_validate(data)
+        return r
+
+    async def update_emergency_call_notification(self, location_id: str, setting: LocationEmergencyCallNotification,
+                                           org_id: str = None):
+        """
+        Update a location emergency call notification.
+
+        Once settings enabled at the organization level, the configured email address will receive emergency call
+        notifications for all locations; for specific location customization, users can navigate to Management >
+        Locations, select the Calling tab, and update the Emergency Call Notification settings.
+
+        Emergency Call Notifications can be enabled at the organization level, allowing specified email addresses to
+        receive email notifications when an emergency call is made. Once activated at the organization level,
+        individual locations can configure this setting to direct notifications to specific email addresses. To comply
+        with U.S. Public Law 115-127, also known as Kari’s Law, any call that's made from within your organization to
+        emergency services must generate an email notification.
+
+        To update location call notification requires a full, user or location administrator auth token with a scope of
+        `spark-admin:telephony_config_write`.
+
+        :param location_id: Update Emergency Call Notification attributes for this location.
+        :type location_id: str
+        :param setting: new settings
+        :type setting: LocationEmergencyCallNotification
+        :param org_id: Update Emergency Call Notification attributes for a location in this organization.
+        :type org_id: str
+        :rtype: None
+        """
+        params = org_id and {'orgId': org_id} or None
+        body = setting.update()
+        url = self.ep(f'{location_id}/emergencyCallNotification')
+        await super().put(url, params=params, json=body)
+
+
 class AsLocationMoHApi(AsApiChild, base='telephony/config/locations'):
     """
     Access codes API
@@ -16731,6 +16995,8 @@ class AsReceptionistContactsDirectoryApi(AsApiChild, base='telephony/config/loca
 
 
 class AsTelephonyLocationApi(AsApiChild, base='telephony/config/locations'):
+    #: emergency services
+    emergency_services: AsLocationEmergencyServicesApi
     #: call intercept settings
     intercept: AsLocationInterceptApi
     #: internal dialing settings
@@ -16746,6 +17012,7 @@ class AsTelephonyLocationApi(AsApiChild, base='telephony/config/locations'):
 
     def __init__(self, session: AsRestSession):
         super().__init__(session=session)
+        self.emergency_services = AsLocationEmergencyServicesApi(session=session)
         self.intercept = AsLocationInterceptApi(session=session)
         self.internal_dialing = AsInternalDialingApi(session=session)
         self.moh = AsLocationMoHApi(session=session)
@@ -16978,6 +17245,8 @@ class AsVirtualLinesApi(AsApiChild, base='telephony/config/virtualLines'):
     call_waiting: AsCallWaitingApi
     #: caller id settings
     caller_id: AsCallerIdApi
+    #: ECBN settings
+    ecbn: AsUserEmergencyServicesApi
     #: forwarding settings
     forwarding: AsPersonForwardingApi
     #: incoming permissions
@@ -16992,6 +17261,7 @@ class AsVirtualLinesApi(AsApiChild, base='telephony/config/virtualLines'):
         self.call_recording = AsCallRecordingApi(session=session, selector=ApiSelector.virtual_line)
         self.call_waiting = AsCallWaitingApi(session=session, selector=ApiSelector.virtual_line)
         self.caller_id = AsCallerIdApi(session=session, selector=ApiSelector.virtual_line)
+        self.ecbn = AsUserEmergencyServicesApi(session=session, selector=ApiSelector.virtual_line)
         self.forwarding = AsPersonForwardingApi(session=session, selector=ApiSelector.virtual_line)
         self.permissions_in = AsIncomingPermissionsApi(session=session, selector=ApiSelector.virtual_line)
         self.permissions_out = AsOutgoingPermissionsApi(session=session, selector=ApiSelector.virtual_line)
@@ -17811,6 +18081,8 @@ class AsTelephonyApi(AsApiChild, base='telephony/config'):
     callqueue: AsCallQueueApi
     call_recording: AsCallRecordingSettingsApi
     dect_devices: AsDECTDevicesApi
+    #: emergency services
+    emergency_services: AsOrgEmergencyServicesApi
     #: WxC device operations
     devices: AsTelephonyDevicesApi
     huntgroup: AsHuntGroupApi
@@ -17846,6 +18118,7 @@ class AsTelephonyApi(AsApiChild, base='telephony/config'):
         self.callqueue = AsCallQueueApi(session=session)
         self.dect_devices = AsDECTDevicesApi(session=session)
         self.devices = AsTelephonyDevicesApi(session=session)
+        self.emergency_services = AsOrgEmergencyServicesApi(session=session)
         self.huntgroup = AsHuntGroupApi(session=session)
         self.jobs = AsJobsApi(session=session)
         self.location = AsTelephonyLocationApi(session=session)
@@ -18754,6 +19027,7 @@ class AsWorkspaceSettingsApi(AsApiChild, base='workspaces'):
     call_waiting: AsCallWaitingApi
     caller_id: AsCallerIdApi
     devices: AsWorkspaceDevicesApi
+    ecbn: AsUserEmergencyServicesApi
     forwarding: AsPersonForwardingApi
     monitoring: AsMonitoringApi
     numbers: AsWorkspaceNumbersApi
@@ -18767,6 +19041,7 @@ class AsWorkspaceSettingsApi(AsApiChild, base='workspaces'):
         self.call_waiting = AsCallWaitingApi(session=session, selector=ApiSelector.workspace)
         self.caller_id = AsCallerIdApi(session=session, selector=ApiSelector.workspace)
         self.devices = AsWorkspaceDevicesApi(session=session)
+        self.ecbn = AsUserEmergencyServicesApi(session=session, selector=ApiSelector.workspace)
         self.forwarding = AsPersonForwardingApi(session=session, selector=ApiSelector.workspace)
         self.monitoring = AsMonitoringApi(session=session, selector=ApiSelector.workspace)
         self.numbers = AsWorkspaceNumbersApi(session=session)